move publish settings
